SimpleSaml symlinks
#### Description
I added CustomHandler script to build symlinks to the SimpleSaml configuration files.

The script will first clean up any old symlinks so there is no conflict. It then builds a symlink to each of the files in the `pritvate/simplesaml` directory. The symlinks are not tracked by GIT and will be recreated anytime that `composer install` or `composer create` are run.